+++ src/main/kotlin/com/windea/plugin/idea/paradox/script/editor/ParadoxDefinitionLocalisationLineMarkerProvider.kt
@@ -9,11 +9,10 @@ import com.intellij.util.*
 import com.windea.plugin.idea.paradox.*
 import com.windea.plugin.idea.paradox.script.psi.*
 
-class ParadoxDefinitionLocalisationLineMarkerProvider: LineMarkerProviderDescriptor(){
+class ParadoxDefinitionLocalisationLineMarkerProvider : LineMarkerProviderDescriptor() {
 	companion object {
 		private val _name = message("paradox.script.gutterIcon.definitionLocalisation")
 		private val _title = message("paradox.script.gutterIcon.definitionLocalisation.title")
-		private fun _tooltip(name: String,type:String) = message("paradox.script.gutterIcon.definitionLocalisation.tooltip", name, type)
 	}
 	
 	override fun getName() = _name
@@ -23,25 +22,35 @@ class ParadoxDefinitionLocalisationLineMarkerProvider: LineMarkerProviderDescrip
 	override fun getLineMarkerInfo(element: PsiElement): LineMarker? {
 		return when(element) {
 			is ParadoxScriptProperty -> {
-				val typeInfo = element.paradoxTypeInfo ?: return null
-				if(!typeInfo.hasLocalisation) return null //没有localisation时不加上gutterIcon
-				LineMarker(element, typeInfo)
+				val definition = element.paradoxDefinition ?: return null
+				if(!definition.hasLocalisation) return null //没有localisation时不加上gutterIcon
+				LineMarker(element, definition)
 			}
 			else -> null
 		}
 	}
 	
-	class LineMarker(element: ParadoxScriptProperty,typeInfo: ParadoxTypeInfo) : LineMarkerInfo<PsiElement>(
+	class LineMarker(element: ParadoxScriptProperty, definition: ParadoxDefinition) : LineMarkerInfo<PsiElement>(
 		element.propertyKey.let { it.propertyKeyId ?: it.quotedPropertyKeyId!! },
 		element.textRange,
 		definitionLocalisationGutterIcon,
-		{ typeInfo.localisation.joinToString("<br>"){ (k,v)-> _tooltip(v,k.value) }},
+		{
+			buildString {
+				val localisation = definition.localisation
+				var isFirst = true
+				for((k, v) in localisation) {
+					if(isFirst) isFirst = false else appendBr()
+					append("(definition localisation) ").append(k.value).append(" = <b>").appendPsiLink("#", v).append("</b>")
+				}
+			}
+		},
 		{ mouseEvent, _ ->
-			val names = typeInfo.localisationValueKeys
+			val names = definition.localisationValueKeys
 			val project = element.project
-			val elements = findLocalisations(names, null, project,hasDefault=true,keepOrder= true).toTypedArray()
+			val elements = findLocalisations(names, null, project, hasDefault = true, keepOrder = true).toTypedArray()
 			when(elements.size) {
-				0 -> {}
+				0 -> {
+				}
 				1 -> OpenSourceUtil.navigate(true, elements.first())
 				else -> NavigationUtil.getPsiElementPopup(elements, _title).show(RelativePoint(mouseEvent))
 			}